Remove the ancient file comment of straight-line.lisp
Ten months ago, I try to implement the Tiger language in Common Lisp,
it ends up I only finish the parsing part of the Tiger language.
(It felt tedious to handle many syntax cases at that time.)
During that time, I license the repository under the GPL licence,
but this time, I simply license the repository under the MIT license,
so the license comment at the beginning of straight-line.lisp
is actually wrong, I simply forgot to remove it when I copy the
(cl:defpackage ...) and (cl:in-package ...) forms.