Separate type checking and IR translation even that need  code duplication
Separate type checking from other tasks, so that find-and-fill-escapes
and translate-program don't need to care about type and other semantic
errors (like undefined variables), and the code is easier to read.

It seems not easy to come up with an easy to use abstraction to
establish type and value environments for both type checking and IR
translation, the mutual recursive calling structure and different
demand for arguments and return values make it hard to abstract a good
pattern, so I decide to just duplicate some code.